manifold: keep the metric layer free of coordinates
The metric layer was importing the coordinate layer: `Geometry for MeshCoords`
sat in `metric::geometry`, and `SimplexLengths::from_coords` in
`metric::simplex`. That is the intrinsic module depending on the extrinsic one,
which inverts invariant 2 -- an embedding induces a metric, a metric induces no
embedding.

Both move to the `coord` side of the boundary, where the dependency runs
downward: `MeshCoords` knows about the metric it induces, and the metric layer
now knows nothing of coordinates at all.

Part of #85.
